+++ NEWS.md
@@ -43,6 +43,7 @@
 - Define Monograph profile for Instance [MODLD-825](https://folio-org.atlassian.net/browse/MODLD-825)
 - Update admin metadata created date handling [MODLD-827](https://folio-org.atlassian.net/browse/MODLD-827)
 - Replace ID_EAN with ID_IAN. Remove property EAN_VALUE [MODLD-810](https://folio-org.atlassian.net/browse/MODLD-810)
+- Update API to support Administrative Metadata [MODLD-819](https://folio-org.atlassian.net/browse/MODLD-819)
 
 ## 1.0.4 (04-24-2025)
 - Work Edit form - Instance read-only section: "Notes about the instance" data is not shown [MODLD-716](https://folio-org.atlassian.net/browse/MODLD-716)